requirements.txt
asgiref==3.7.2
Django==4.2.1
django-grappelli==3.0.8
django-jazzmin==2.6.0
django-modeltranslation==0.18.11
pillow==10.2.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.6.3
